requirements.txt
# Data Processing Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0

# Data Extraction Libraries
requests>=2.25.0
beautifulsoup4>=4.9.0

# Network Optimization Libraries
tensorflow>=2.8.0
torch>=1.11.0

# Energy Efficiency Optimization Libraries
pulp>=2.4

# Predictive Network Planning Libraries
scipy>=1.7.0
statsmodels>=0.12.0

# Network Anomaly Detection Libraries
matplotlib>=3.4.0
seaborn>=0.11.0
pyod>=0.9.0

# Visualization Libraries
plotly>=5.0.0

# Configuration and Utilities
PyYAML>=5.4.0
joblib>=1.0.0

# Testing
pytest>=6.2.0

# Note: This file lists all the required libraries and their versions for the Network Optimization System.
# To install these libraries, run the following command:
# pip install -r requirements.txt
# Note: It's recommended to create a new virtual environment before installing these libraries to avoid conflicts with existing libraries on your system.
